refactor(db): move sqlite ORM wrappers to their bounded contexts
EdoProSQLiteTypeORM is edopro-only (used by bootstrapPersistence) and
PreReleasesMercurySQLiteTypeORM is ygopro-only (used by the mercury prereleases
util), yet both lived under src/shared/db. Move them to
edopro/card/infrastructure/sqlite and ygopro/card/infrastructure/sqlite
respectively. Only data-source.ts and the shared card entities remain under
shared, which is what both contexts genuinely share.
